Refactoring/commenting
No big changes

+++ time4io.c
@@ -2,8 +2,8 @@
 #include <pic32mx.h>
 #include "mipslab.h"
 
-// Everything below is written by Theodor Björkman
-int getsw(void)
+// Everything below is written by Theodor Björkman. Latest modification is 2023-03-01
+uint8_t getsw(void)
 {
     int r = 0xf00;
     r &= PORTD;
@@ -11,7 +11,7 @@ int getsw(void)
     return r;
 }
 
-int getbtns(void)
+uint8_t getbtns(void)
 {
     int r = 0xe0;
     r &= PORTD;
